Version Panel & Dashboard & Overall code cleanup
I fixed all the issues I previously made. But created a lot of new issues when trying to implement a new system to update the server Version file. Found a new solution to fix this issue, but no time left.

Give link in the input box of the version modal. When presses on the button, the link will be send to the socket server, who sends it forward to the Minecraft server. De Minecraft server will then surf to the URL and download the file. When downloaded, the current version file will be replaced with a new one and a success message will be send to the web-client.